requirements.txt
LenHTTP == 2.2.7
aioredis == 1.3.0
bcrypt == 3.2.0
aiomysql == 0.0.21
aiohttp == 3.7.4.post0
peace-performance-python == 1.1.2
py3rijndael == 0.3.3
colorama == 0.4.4
tabulate == 0.8.9
progressbar2 == 3.55.0
aiopath == 0.6.10
# OPTIONALS which may improve speed but lower compatibility.
orjson == 3.6.3
uvloop == 0.16.0
osupyparser == 0.2.0
